Add hotswap method to Lua and VM for live source code updates

+++ tests/multi.rs
@@ -12,7 +12,7 @@ fn multiple_returns() {
         "#;
 
     if let ExVal::Table(t) = simple(source_in) {
-        assert_eq!(t.get("a"), Some(&ExVal::Integer(1)));
+        assert_eq!(t.get("a"), Some(&ExVal::Integer(2)));
         assert_eq!(t.get("b"), Some(&ExVal::String("hello".to_string())));
         assert_eq!(t.get("c"), Some(&ExVal::Bool(true)));
     } else {
